Use standard declaration of real variables using "selected_real_kind".

+++ src/approximation/bsplvd.f90
@@ -30,9 +30,9 @@ SUBROUTINE BSPLVD(T,K,X,Ileft,Vnikx,Nderiv)
   !   900328  Added TYPE section.  (WRB)
 
   INTEGER :: Ileft, K, Nderiv
-  REAL :: T(:), Vnikx(K,K), X
+  REAL(SP) :: T(:), Vnikx(K,K), X
   INTEGER :: i, ideriv, idervm, ipkmd, j, jlow, jm1, kmd, l, m
-  REAL :: a(20,20), diff, fkmd, v
+  REAL(SP) :: a(20,20), diff, fkmd, v
   !* FIRST EXECUTABLE STATEMENT  BSPLVD
   CALL BSPLVN(T,K+1-Nderiv,1,X,Ileft,Vnikx(Nderiv,Nderiv))
   IF ( Nderiv>1 ) THEN
