+++ src/approximation/cv.f90
@@ -1,5 +1,5 @@
 !** CV
-REAL FUNCTION CV(Xval,Ndata,Nconst,Nord,Nbkpt,Bkpt,W)
+REAL(SP) FUNCTION CV(Xval,Ndata,Nconst,Nord,Nbkpt,Bkpt,W)
   !>
   !  Evaluate the variance function of the curve obtained
   !            by the constrained B-spline fitting subprogram FC.
@@ -106,9 +106,9 @@ REAL FUNCTION CV(Xval,Ndata,Nconst,Nord,Nbkpt,Bkpt,W)
   !   891214  Prologue converted to Version 4.0 format.  (BAB)
   !   920501  Reformatted the REFERENCES section.  (WRB)
   INTEGER :: Nbkpt, Nconst, Ndata, Nord
-  REAL :: Xval, Bkpt(Nbkpt), W(:)
+  REAL(SP) :: Xval, Bkpt(Nbkpt), W(:)
   INTEGER :: i, ileft, ip, is, last, mdg, mdw, n
-  REAL :: zero, v(40)
+  REAL(SP) :: zero, v(40)
   !* FIRST EXECUTABLE STATEMENT  CV
   zero = 0.
   mdg = Nbkpt - Nord + 3
