+++ src/approximation/dbols.f90
@@ -435,10 +435,10 @@ SUBROUTINE DBOLS(W,Mdw,Mrows,Ncols,Bl,Bu,Ind,Iopt,X,Rnorm,Mode,Rw,Iw)
   USE linear, ONLY : DROT, DROTG
   INTEGER :: Mdw, Mode, Mrows, Ncols
   INTEGER :: Ind(Ncols), Iw(2*Ncols), Iopt(*)
-  REAL(8) :: Rnorm, W(Mdw,Ncols+1), Bl(Ncols), Bu(Ncols), X(2*Ncols), Rw(5*Ncols)
+  REAL(DP) :: Rnorm, W(Mdw,Ncols+1), Bl(Ncols), Bu(Ncols), X(2*Ncols), Rw(5*Ncols)
   INTEGER :: i, ibig, inrows, ip, j, jp, lds, lenx, liopt, llb, lliw, llrw, &
     llx, lmdw, lndw, locdim, lp,mnew, nerr
-  REAL(8) :: sc, ss, one, zero
+  REAL(DP) :: sc, ss, one, zero
   LOGICAL checkl
   CHARACTER(8) :: xern1, xern2
   CHARACTER(16) :: xern3, xern4
