+++ src/approximation/dmout.f90
@@ -64,7 +64,7 @@ SUBROUTINE DMOUT(M,N,Lda,A,Ifmt,Idigit)
   !   910403  Updated AUTHOR section.  (WRB)
   USE service, ONLY : I1MACH
   INTEGER :: Idigit, Lda, M, N
-  REAL(8) :: A(Lda,N)
+  REAL(DP) :: A(Lda,N)
   CHARACTER(*) Ifmt
   INTEGER :: i, j, k1, k2, lout, ndigit
   CHARACTER(3), PARAMETER :: icol = 'COL'
