+++ src/approximation/fc.f90
@@ -381,7 +381,7 @@ SUBROUTINE FC(Ndata,Xdata,Ydata,Sddata,Nord,Nbkpt,Bkpt,Nconst,Xconst,&
   !   920501  Reformatted the REFERENCES section.  (WRB)
   USE service, ONLY : XERMSG
   INTEGER :: Mode, Nbkpt, Nconst, Ndata, Nord, Nderiv(Nconst), Iw(:)
-  REAL :: Bkpt(Nbkpt), Coeff(:), Sddata(Ndata), W(*), Xconst(Nconst), Xdata(Ndata), &
+  REAL(SP) :: Bkpt(Nbkpt), Coeff(:), Sddata(Ndata), W(*), Xconst(Nconst), Xdata(Ndata), &
     Yconst(Nconst), Ydata(Ndata)
   !
   INTEGER :: i1, i2, i3, i4, i5, i6, i7, mdg, mdw
